clean up component/split out utility functions
